backend/mcp-server/requirements.txt
# MCP Server Dependencies

# FastMCP for building MCP servers
fastmcp==1.2.0

# MCP SDK
mcp==1.26.0

# Database
sqlalchemy==2.0.36

# Claude API for AI narratives
anthropic==0.42.0

# HTTP client
httpx==0.28.1

# Data manipulation
pandas==2.2.3
numpy==2.2.2

# Utilities
python-dotenv==1.0.1
pydantic==2.10.4
